Dockerfile: DRY the cross-build architecture list
The list of cross-build architectures was repeated in several places in
the Dockerfile. Extract the list to a build argument and use bash brace
expansions to interpolate the list into command arguments. Bash performs
brace expansion before variable expansion so eval in some form is
required to make this scheme work. Furthermore, brace expansion is a
bashism and the shell for RUN commands (/bin/sh) does not support them.
As switching the shell for the whole Dockerfile carries some risk of
breaking things and variable-expanded script fragments need to be
evaluated regardless, I elected to instead take the low-risk---albeit
slightly inefficient---approach of invoking bash in a subshell to
perform the brace expansions.

Signed-off-by: Cory Snider <csnider@mirantis.com>