logger/journald: format cgo
When editing this file in my IDE, it re-formatted some lines that did
not have a leading space.

I'm not 100% sure how whitespace is treated in these, but only changing
"some" lines to have a space, and others to only have a TAB as indentation
didn't seem correct. As the existing `#include ..` lines already had a
leading space, I assumed adding a space to all of the lines would be ok,
so I manually updated all remaining lines to have a leading space.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>