Filter images, not their summaries
The List Images API endpoint's `filters` parameter can be a bit quirky.
Unlike most of the filter predicates, which are applied by including or
excluding each image from the list, the `reference` predicate filters
out individual references from the image summary's `RepoTags` and
`RepoDigests` fields. This behaviour is surprising to our users as they
cannot use the API to look up the repo digests for a specified image tag
(or vice versa) without widening the filter to match on only the
repository and post-filtering the results. Change the API such that the
filters are only used to determine whether or not to include a
particular image in the list of image summaries, without filtering the
contents of the summary.

Fixes #43830

Signed-off-by: Cory Snider <csnider@mirantis.com>