Dockerfile: handle multi base image for riscv64 and armel support
To add support for riscv64 builds we need crossbuild packages for
riscv64 but current golang image with debian bullseye does not
support it. Ubuntu 22.04 supports riscv64 but unfortunately drops
support for armel arch. Therefore we need a multi base image that
will be picked up based on the target platform we want to build.

Signed-off-by: CrazyMax <crazy-max@users.noreply.github.com>