integration-cli: rewrite TestBuildAPIDockerFileRemote without "find"
This test was originally added in 15924f238503ffe44e5fcb99415ff753a36e5971 to
verify that the custom Dockerfile name did not end up in the image. The test was
using `find` for this, to search inside the image's `/tmp` directory but after
an update in a853f8e4687adb78c0e63019caae32b216cce925 (likely because the
Dockerfile would now be included at the root, because Windows may not have `tmp`),
if was now searching the whole image's rootfs. This is a bit of a heavy search,
and traversing the whole rootfs could lead to EPERM (f453261b4a53ab71db8b78ea119b4edd69b95ae3)
issues on Linux, and (with newer Busybox versions) on Windows;

    The command 'cmd /S /C find / -name ba*' returned a non-zero code: 1
    ...
    find: /Program Files/Windows Defender Advanced Threat Protection/Classification/Configuration: Permission denied
    find: /ProgramData/Microsoft/Diagnosis/feedbackhub: Permission denied
    find: /ProgramData/Microsoft/Diagnosis/temp: Permission denied
    find: /ProgramData/Microsoft/Diagnosis/timetraveldebuggingstorage: Permission denied

This patch updates the test to `stat` in the expected locations (`/` and `/tmp`),
to make sure the `baz` file isn't found.

With this patch:

    make BIND_DIR=. DOCKER_GRAPHDRIVER=vfs TEST_FILTER=TestBuildAPIDockerFileRemote test-integration
    ...
    === RUN   TestDockerSuite/TestBuildAPIDockerFileRemote
      --- PASS: TestDockerSuite (3.04s)
        --- PASS: TestDockerSuite/TestBuildAPIDockerFileRemote (3.02s)

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>