client: disable deprecated NewClient() and NewEnvClient()
The `NewClient()` and `NewEnvClient()` functions were deprecated in 18.03 in
commit 772edd020cc784913973387b00c4d0c526a10a26, but they continued to be functional.

Searching public repositories on GitHub shows that there's quite some projects
that haven't updated their code to use the replacements.

This patch updates both functions to become non-functional (to enforce a compilation
failure), but extend the function's documentaiton with instructions on migrating to
their non-deprecated alternatives. This should hopefully assist projects to migrate
their code.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>