Dockerfile: containerd cross
containerd build in Dockerfile is limited to
host platform and could not be cross-built
for other platforms.

this change allows to build against any platforms
if we want to smoke test in a follow-up but also
enhance e2e tests for linux and windows in our
pipeline.

also introduced DOCKER_LINKMODE to be able to
build dynamic or static binaries.

Signed-off-by: CrazyMax <crazy-max@users.noreply.github.com>