pkg/archive: adjust permissions for parent directories
Use 0755 by default if the tar archive does not explicitly provide the
mode settings for parent directories.

This change makes the permissions and ownership of the resulting
directory aligned with the implementation of the Upack function from
pkg/archive/archive.go. Previously this was causing inconsistencies:
depending on the configured graphdriver some directories were created
with 0600 access bits.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>