daemon: fix mount loop on 'docker cp'
On modern Linux systems, the propagation mode for all mounts in the
initial namespace is set to shared early in boot. The propagation mode
of any container root directory is therefore also set to shared on such
systems as the default propagation mode of any mount is inherited from
the propagation mode of its parent mount. The runtime changes the
propagation mode of the container namespace's '/' mount to slave (by
default) to prevent mounts inside the container from propagating out
into the initial namespace, but any filesystems mounted under the rootfs
in the initial namespace could still propagate into the container.

When a 'docker cp' command is invoked, the daemon mounts all container
volumes under the rootfs directory in the initial namespace so that
files can be copied into and out of paths located inside volumes. The
mounts propagate into the container's namespace, overmounting the volume
mounts with themselves. This is not problematic in trivial cases as the
unmounts also propagate into the container namespace, reverting the
overmounts. However, when a parent directory of the container's rootfs
is configured as a container volume, the mounts propagated into the
container namespace for 'docker cp' are not fully reversed when the
daemon unmounts them in the initial namespace. Subsequent commands
propagate new mounts into the container namespace which do not get
cleaned up, which build up until the mount table fills to its limit and
subsequent mount operations fail with ENOSPC.

Set the propagation mode for the container rootfs path to private before
starting the container so that child mounts added in the initial
namespace after the fact never propagate into the container namespace in
the first place.

Signed-off-by: Cory Snider <csnider@mirantis.com>