daemon: replace ErrRootFSReadOnly with errdefs
It was only used in a single location, and the ErrRootFSReadOnly was not checked
for, or used as a sentinel error.

This error was introduced in c32dde5baadc8c472666ef9d5cead13ab6de28ea, originally
named `ErrContainerRootfsReadonly`. It was never used as a sentinel error, but
from that commit, it looks like it was added as a package variable to mirror
the coding style of already existing errors defined at the package level.

This patch removes the exported variable, and replaces the error with an
errdefs.InvalidParameter(), so that the API also returns the correct (400)
status code.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>